requirements.txt
neo4j==5.9.0
azure-mgmt-resource==23.0.1
azure-recommendations-11==0.2.0
azure-mgmt-compute==30.0.0
psycopg2==2.9.6
azure-identity==1.13.0
click==8.1.3
tabulate==0.9.0

